+++ docs/source/user_guide/changelog.rst
@@ -12,6 +12,7 @@ Changelog
 - Return to the penultimate accumulation step if the last step as broken
 - Focus on error when selecting an "No id" error in Validator
 - Improve Validator responsiveness when loading a large session.
+- Improve logging information
 
 5.1.3
 =====